requirements.txt
# Core
anthropic>=0.40.0
python-dotenv>=1.0.0
requests>=2.31.0

# Trends
pytrends>=4.9.2

# Voice — ElevenLabs (replaces edge-tts)
# No extra package needed, uses requests directly

# Video assembly — FFmpeg (system install, no pip package)
# Install: sudo apt install ffmpeg  OR  brew install ffmpeg

# YouTube upload
google-api-python-client>=2.100.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0

# Dashboard
flask>=3.0.0
Pillow>=10.0.0

# Scheduler
APScheduler>=3.10.0
